% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tropicalTongue.R
\name{tropicalTongue}
\alias{tropicalTongue}
\title{Estimate the proboscis length of tropical bees}
\usage{
tropicalTongue(fam, tribe, genus, IT_mm)
}
\arguments{
\item{fam}{A character vector containing a combination of Apidae, Andrenidae, Halictidae, Megachilidae or Melittidae}

\item{tribe}{A character vector of the same length containing the tribes}

\item{genus}{A character vector of the same length containing the genera}

\item{IT_mm}{A numeric vector containing the intertegular distane in mm}
}
\value{
A numeric vector
}
\description{
Estimate the proboscis length of tropical bees
}
\examples{
tropicalTongue("Apidae", "Meliponini", "Trigona", 0.7)

}
