% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tropicalTongue.R
\name{tropicalTongue}
\alias{tropicalTongue}
\title{Estimating the proboscis length of tropical bees}
\usage{
tropicalTongue(fam, tribe, genus, IT_mm)
}
\arguments{
\item{fam}{A character vector indicating the family by containing a combination of Apidae, Andrenidae, Halictidae, Megachilidae or Melittidae}

\item{tribe}{A character vector of the same length containing the tribes}

\item{genus}{A character vector of the same length containing the genera.}

\item{IT_mm}{A numeric vector containing the intertegular distane in mm}
}
\value{
A numeric vector
}
\description{
Estimating the proboscis length of tropical bees
}
\details{
\itemize{
\item Parameter fam accepts the following values: Apidae, Andrenidae, Halictidae, Megachilidae, Melittidae
\item Parameter tribe accepts the following values: Augochlorini, Euglossini, Meliponini
\item Parameter genus accepts the following values: Augochlora, Augochlorella, Augochloropsis, Corynura, Dolichotrigona,
Eufriesa, Euglossa (Glossura), Euglossa (Glossurella)Euglossa sensu stricto, Euglossa, Eulaema, Exarete, Melipona,
Nannotrigona, Nogueirapis, Oxytrigona, Paratrigona, Partamona, Plebeia, Ptilotrigona, Scaptotrigona, Scaura, Tetragona, Tetragonisca, Trigona, Trigonisca
}
}
\examples{
tropicalTongue("Apidae", "Meliponini", "Trigona", 0.7)

}
